requirements.txt
Django==5.0.6
django-allauth==0.63.3
django-environ==0.11.2
psycopg2-binary==2.9.9
stripe==10.4.0
Pillow==10.4.0
whitenoise==6.7.0
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.1
